requirements.txt
faker==37.1.0
pandas==2.2.3
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
python-dotenv==1.0.1
apache-airflow==2.10.5; python_version < "3.13"